Get altoclef to go back for furnace
Gets altoclef to go back for furnace after initially smelting iron gear. This prevents the current situation where it gets a lot of food by walking everywhere and has to walk a very long distance back before it can cook that food at the same furnace.

This commit also makes altoclef go back for the furnace after the food itself is cooked. Otherwise, it would leave the furnace in place, collect diamond ore, and then again travel the very long distance back to be able to re-smelt that diamond ore in the same furnace it cooked the food in.

This commit attempts to help cut down on these repetitive trips.

Seems to work.

Future: (Maybe?) Just craft two or three furnaces to begin with? So that we can just leave them behind. Same for crafting tables.